refactor: move TestFile into util module and simplify imports
