+++ src/lib.rs
@@ -4,6 +4,4 @@ pub mod cli;
 pub mod disk;
 pub mod page;
 pub mod storage;
-
-#[cfg(test)]
-pub(crate) mod test_utils;
+pub mod util;
