+++ src/storage/file_storage.rs
@@ -123,7 +123,7 @@ impl StorageEngine for FileStorage {
 mod tests {
     use super::*;
     use crate::storage::error::StorageError;
-    use crate::test_utils::TestFile;
+    use crate::util::TestFile;
 
     /// This test verifies that opening a file with `FileStorage::open` creates the file if it does not already exist. It uses the `TestFile` struct to generate a unique temporary file path, attempts to open the file using `FileStorage`, and then checks that the operation was successful and that the file now exists on the filesystem.
     #[test]
