+++ tests/access_to_storage_integration_test.rs
@@ -3,11 +3,8 @@ use mini_rdbms_rs::buffer::LruBufferManager;
 use mini_rdbms_rs::disk::{PageId, StorageDiskManager};
 use mini_rdbms_rs::page::{PageError, SlottedPageManager};
 use mini_rdbms_rs::storage::FileStorage;
+use mini_rdbms_rs::util::TestFile;
 use std::path::Path;
-
-#[path = "../src/test_utils.rs"]
-mod test_utils;
-use test_utils::TestFile;
 
 type TestDiskManager = StorageDiskManager<FileStorage>;
 type TestBufferManager = LruBufferManager<TestDiskManager>;
